Added a Disclaimer
Disclaimer -
This project is no longer a casual student project.
It is too complex and featurefull for a basic project.
Keep in mind that this is a project by a SOLO 16 year old developer.
Please DO NOT STEAL the encryption/Decryption algorithm as for a small period of time, this code is public..